test,data:Add tests for Value::push_back(). refactor:Types.[cpp|h], RoundTripTest.cpp:Rename PortType::TYPE_VEC3D to TYPE_VEC2 to reflect available vector of floats. opt:Value.h:push_back():Remove switch on type in favour of copying ourselves into the vector.

+++ src/tests/DagTest/DagTest.cpp
@@ -74,8 +74,8 @@ TEST(MemoryNodeLibraryTest_testClassNotFound, checkClassNotFound)
 TEST(ValueTest_testIncrement, checkValueChanges)
 {
     dagbase::Value sut;
-    sut += std::int64_t(1);
-    EXPECT_EQ(1, std::int64_t(sut));
+    sut += std::uint8_t(1);
+    EXPECT_EQ(1, std::uint8_t(sut));
 }
 
 class NodeCategoryTest : public ::testing::TestWithParam<std::tuple<const char*, const char*, dagbase::NodeCategory::Category>>
